requirements.txt
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
distlib==0.3.7
filelock==3.12.4
future==0.18.3
geocoder==1.38.1
identify==2.5.30
idna==3.4
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
ratelim==0.1.6
requests==2.31.0
six==1.16.0
urllib3==2.0.6
virtualenv==20.24.5
